--- data/src/main/java/com/openbank/data/data/entity/Comics.kt
@@ -1,14 +1,0 @@
-package com.openbank.data.data.entity
-
-
-class Comics {
-
-    var available: Long = 0
-
-    var collectionURI: String? = null
-
-    var items: List<Item>? =
-        null
-    var returned: Long = 0
-
-}
