--- data/src/main/java/com/openbank/data/data/entity/Item.kt
@@ -1,8 +1,0 @@
-package com.openbank.data.data.entity
-
-
-class Item {
-
-    var resourceURI: String? = null
-    var name: String? = null
-}
