--- data/src/main/java/com/openbank/data/data/entity/MarvelCharactersResponse.kt
@@ -1,20 +1,0 @@
-package com.openbank.data.data.entity
-
-
-class MarvelCharactersResponse {
-
-    var code: Long = 0
-
-    var status: String? = null
-
-    var copyright: String? = null
-
-    var attributionText: String? = null
-
-    var attributionHTML: String? = null
-
-    var etag: String? = null
-
-    var data: Data? = null
-
-}
