--- data/src/main/java/com/openbank/data/data/entity/Series.kt
@@ -1,13 +1,0 @@
-package com.openbank.data.data.entity
-
-class Series {
-
-    var available: Long = 0
-    var collectionURI: String? = null
-
-    var items: List<Item>? =
-        null
-
-    var returned: Long = 0
-
-}
