--- data/src/main/java/com/openbank/data/data/entity/Stories.kt
@@ -1,10 +1,0 @@
-package com.openbank.data.data.entity
-
-class Stories {
-
-    var available: Long = 0
-    var collectionURI: String? = null
-    var storiesItems: List<StoriesItem>? = null
-    var returned: Long = 0
-
-}
