--- data/src/main/java/com/openbank/data/data/entity/StoriesItem.kt
@@ -1,10 +1,0 @@
-package com.openbank.data.data.entity
-
-class StoriesItem {
-
-    var resourceURI: String? = null
-
-    var name: String? = null
-    var type: String? = null
-
-}
