--- data/src/main/java/com/openbank/data/data/entity/Thumbnail.kt
@@ -1,8 +1,0 @@
-package com.openbank.data.data.entity
-
-class Thumbnail {
-
-    var path: String? = null
-    var extension: String? = null
-
-}
