--- data/src/main/java/com/openbank/data/data/entity/Url.kt
@@ -1,9 +1,0 @@
-package com.openbank.data.data.entity
-
-class Url {
-
-    var type: String? = null
-
-    var url: String? = null
-
-}
