--- data/src/main/java/com/openbank/data/data/entity/test.kt
@@ -1,4 +1,0 @@
-package com.openbank.data.data.entity
-
-class test {
-}
