--- data/src/main/java/com/openbank/data/data/source/remote/GetCharacterDetailsRemoteSource.kt
@@ -1,10 +1,0 @@
-package com.openbank.data.data.source.remote
-
-import com.openbank.domain.uimodel.CharacterDetailsUiModel
-import com.openbank.domain.uimodel.CharacterListUiModel
-import io.reactivex.Observable
-
-interface GetCharacterDetailsRemoteSource {
-
- fun getMarvelCharacterDetails(characterId:Long): Observable<List<CharacterDetailsUiModel>>
-}
