--- data/src/main/java/com/openbank/data/data/source/remote/GetCharacterListRemoteSource.kt
@@ -1,9 +1,0 @@
-package com.openbank.data.data.source.remote
-
-import com.openbank.domain.uimodel.CharacterListUiModel
-import io.reactivex.Observable
-
-interface GetCharacterListRemoteSource {
-
- fun getMarvelCharacterList(offset:Int): Observable<List<CharacterListUiModel>>
-}
