+++ data/src/main/java/com/openbank/data/entity/Data.kt
@@ -1,0 +1,14 @@
+package com.openbank.data.entity
+
+
+class Data {
+
+    var offset: Long = 0
+    var limit: Long = 0
+
+    var total: Long = 0
+    var count: Long = 0
+    var results: List<Result>? =
+        null
+
+}
