+++ data/src/main/java/com/openbank/data/entity/Events.kt
@@ -1,0 +1,12 @@
+package com.openbank.data.entity
+
+
+
+class Events {
+
+    var available: Long = 0
+    var collectionURI: String? = null
+    var items: List<Item>? = null
+    var returned: Long = 0
+
+}
