+++ data/src/main/java/com/openbank/data/entity/Item.kt
@@ -1,0 +1,8 @@
+package com.openbank.data.entity
+
+
+class Item {
+
+    var resourceURI: String? = null
+    var name: String? = null
+}
