+++ data/src/main/java/com/openbank/data/entity/Series.kt
@@ -1,0 +1,13 @@
+package com.openbank.data.entity
+
+class Series {
+
+    var available: Long = 0
+    var collectionURI: String? = null
+
+    var items: List<Item>? =
+        null
+
+    var returned: Long = 0
+
+}
