+++ data/src/main/java/com/openbank/data/entity/Stories.kt
@@ -1,0 +1,10 @@
+package com.openbank.data.entity
+
+class Stories {
+
+    var available: Long = 0
+    var collectionURI: String? = null
+    var storiesItems: List<StoriesItem>? = null
+    var returned: Long = 0
+
+}
