+++ data/src/main/java/com/openbank/data/entity/StoriesItem.kt
@@ -1,0 +1,10 @@
+package com.openbank.data.entity
+
+class StoriesItem {
+
+    var resourceURI: String? = null
+
+    var name: String? = null
+    var type: String? = null
+
+}
