+++ data/src/main/java/com/openbank/data/entity/Thumbnail.kt
@@ -1,0 +1,8 @@
+package com.openbank.data.entity
+
+class Thumbnail {
+
+    var path: String? = null
+    var extension: String? = null
+
+}
