+++ data/src/main/java/com/openbank/data/entity/Url.kt
@@ -1,0 +1,9 @@
+package com.openbank.data.entity
+
+class Url {
+
+    var type: String? = null
+
+    var url: String? = null
+
+}
