+++ data/src/main/java/com/openbank/data/entity/test.kt
@@ -1,0 +1,4 @@
+package com.openbank.data.entity
+
+class test {
+}
