+++ data/src/main/java/com/openbank/data/mapper/CharacterListMapperImpl.kt
@@ -1,0 +1,28 @@
+package com.openbank.data.mapper
+
+import com.openbank.data.entity.MarvelCharactersResponse
+
+import com.openbank.domain.uimodel.CharacterListUiModel
+import javax.inject.Inject
+
+class CharacterListMapperImpl @Inject constructor():
+    CharacterListMapper {
+
+    override fun getCharactersListUiModel(marvelCharactersResponse: MarvelCharactersResponse):List<CharacterListUiModel>{
+        var characterListMapperArray = ArrayList<CharacterListUiModel>()
+        if (marvelCharactersResponse.data?.results?.size!! > 0) {
+            for (result in marvelCharactersResponse.data?.results!!) {
+                var characterListMapper =
+                   CharacterListUiModel()
+                characterListMapper.characterName = result.name
+                characterListMapper.characterId = result.id
+                characterListMapper.characterDescription = result.description
+                characterListMapper.characterUrl =
+                    result?.thumbnail?.path + "." + result?.thumbnail?.extension
+
+                characterListMapperArray.add(characterListMapper)
+            }
+        }
+        return characterListMapperArray
+    }
+}
