+++ data/src/main/java/com/openbank/data/utils/DateHelper.kt
@@ -1,0 +1,14 @@
+package com.openbank.data.utils
+
+import java.text.SimpleDateFormat
+import java.util.*
+
+object DateHelper {
+
+    fun getCurrentTimestamp(): String? {
+        return SimpleDateFormat("yyyy-MM-dd HH:mm").format(
+            Calendar
+                .getInstance().time
+        )
+    }
+}
