+++ domain/src/main/java/com/openbank/domain/uimodel/CharacterListUiModel.kt
@@ -1,10 +1,9 @@
 package com.openbank.domain.uimodel
 
 
-class CharacterListUiModel {
-    var characterName: String? = null
-    var characterDescription: String? = null
-    var characterUrl: String? = null
+data class CharacterListUiModel (
+    var characterName: String? = null,
+    var characterDescription: String? = null,
+    var characterUrl: String? = null,
     var characterId: Long? = null
-
-}
+)
